# openapi.json
{
 "openapi":"3.0.0",
 "info":{
    "title":"tasks API",
    "version":"1.0.0",
    "description":"A simple CRUD API for managing tasks"
 },
 "paths":{
    "/tasks":{
        "get":{
            "summary":"Get all tasks",
        "responses":{
            "200":{
                "description":"a list of all tasks"
            }
        }
        },
        "post":{
            "summary":"create a new task",
            "requestBody":{
                "required":true,
                "content":{
                    "application/json":{
                        "schema":{
                            "type":"object",
                            "properties":{
                                "title":{
                                    "type":"string"
                                }   
                            },
                            "required":["title"]
                        }
                    }
                }
            },
            "responses":{
                "201":{
                    "description":"task created successfully"
                },
                "400":{
                    "description":"missing required fields"
                }
            }
        }
    },
    "/tasks/{id}":{
        "get":{
            "summary":"get a single task by id",
            "parameters":[
            {"name":"id" , "in": "path" ,"required":true,"schema":{"type": "integer"}}
            ],
            "responses":{
                "200":{"description":"task found"},
                "404":{"description":"task not found"}
            }
        },
        "put":{
            "summary":"update a task",
            "parameters":[
                {"name":"id","in":"path","required":true,"schema":{"type":"integer"}}
            ],
            "requestBody":{
                "required":"true",
                "content":{
                    "application/json":{
                        "schema":{
                            "properties":{
                                "title":{"type":"string"},
                                "done":{"type":"boolean"}
                            },
                            "required":["title"]
                        }
                    }
                }
            },
            "responses":{
                "200":{"description":"updated successfully"},
                "400":{"description":"missing required fields"},
                "404":{"description":"task not found"}
            }
        },
        "delete":{
            "summary":"delete a task",
            "parameters":[
                {"name":"id","in":"path","required":true,"schema":{"type":"integer"}}
            ],
            "responses":{
                "204":{"description":"task deleted successfully"},
                "404":{"description":"task not found"}
            }
        }
    }
 }
}